Desacopla preferência de tema das configurações administrativas
