+++ frontend/src/components/__tests__/MainNavbar.spec.ts
@@ -4,10 +4,10 @@ import {ref} from "vue";
 import {getCommonMountOptions, setupComponentTest} from "@/test-utils/componentTestHelpers";
 import NavBar from "../layout/MainNavbar.vue";
 import {usePerfil} from "@/composables/usePerfil";
-import {useConfiguracoes} from "@/composables/useConfiguracoes";
+import {useTemaPreferencia} from "@/composables/useTemaPreferencia";
 
 vi.mock("@/composables/usePerfil");
-vi.mock("@/composables/useConfiguracoes");
+vi.mock("@/composables/useTemaPreferencia");
 vi.mock("@/services/usuarioService", () => ({
     logout: vi.fn().mockResolvedValue(undefined),
 }));
@@ -36,7 +36,7 @@ describe("MainNavbar.vue", () => {
 
     beforeEach(() => {
         vi.clearAllMocks();
-        vi.mocked(useConfiguracoes).mockReturnValue({
+        vi.mocked(useTemaPreferencia).mockReturnValue({
             getTemaEscuro: vi.fn().mockReturnValue(false),
             setTemaEscuro: vi.fn(),
         } as any);
