+++ frontend/src/composables/__tests__/useConfiguracoes.spec.ts
@@ -107,20 +107,5 @@ describe('useConfiguracoes', () => {
         expect(composable.getDiasAlertaNovo()).toBe(3);
     });
 
-    it('deve persistir tema escuro por usuário no localStorage', () => {
-        const composable = useConfiguracoes();
 
-        composable.setContextoUsuarioTemaEscuro('101');
-        expect(composable.getTemaEscuro()).toBe(false);
-
-        composable.setTemaEscuro(true);
-        expect(composable.getTemaEscuro()).toBe(true);
-
-        composable.setContextoUsuarioTemaEscuro('202');
-        expect(composable.getTemaEscuro()).toBe(false);
-
-        composable.setTemaEscuro(true);
-        composable.setContextoUsuarioTemaEscuro('101');
-        expect(composable.getTemaEscuro()).toBe(true);
-    });
 });
