+++ frontend/src/composables/useConfiguracoes.ts
@@ -9,22 +9,6 @@ import {
 export type {Parametro};
 
 const configuracoes = ref<Parametro[]>([]);
-const CHAVE_TEMA_ESCURO = "temaEscuro";
-const temaEscuro = ref(false);
-const codigoUsuarioTema = ref<string | null>(null);
-
-function montarChaveTemaEscuro(codigoUsuario: string): string {
-    return `${CHAVE_TEMA_ESCURO}:${codigoUsuario}`;
-}
-
-function lerTemaEscuroPorUsuario(codigoUsuario: string): boolean {
-    const valor = localStorage.getItem(montarChaveTemaEscuro(codigoUsuario));
-    return valor === "true";
-}
-
-function salvarTemaEscuroPorUsuario(codigoUsuario: string, novoValor: boolean) {
-    localStorage.setItem(montarChaveTemaEscuro(codigoUsuario), String(novoValor));
-}
 
 export function useConfiguracoes() {
     const {carregando, erro, executar} = useAsyncAction();
@@ -69,27 +53,6 @@ export function useConfiguracoes() {
         return parseInt(valor, 10) || 3;
     }
 
-    function setContextoUsuarioTemaEscuro(codigoUsuario: string | null | undefined) {
-        codigoUsuarioTema.value = codigoUsuario ? String(codigoUsuario) : null;
-        if (!codigoUsuarioTema.value) {
-            temaEscuro.value = false;
-            return;
-        }
-
-        temaEscuro.value = lerTemaEscuroPorUsuario(codigoUsuarioTema.value);
-    }
-
-    function getTemaEscuro(): boolean {
-        return temaEscuro.value;
-    }
-
-    function setTemaEscuro(novoValor: boolean) {
-        temaEscuro.value = novoValor;
-        if (codigoUsuarioTema.value) {
-            salvarTemaEscuroPorUsuario(codigoUsuarioTema.value, novoValor);
-        }
-    }
-
     return {
         configuracoes,
         loading: carregandoConfiguracoes,
@@ -100,8 +63,5 @@ export function useConfiguracoes() {
         getValor,
         getDiasInativacaoProcesso,
         getDiasAlertaNovo,
-        setContextoUsuarioTemaEscuro,
-        getTemaEscuro,
-        setTemaEscuro,
     };
 }
